+++ docs/getting-started/changelog.rst
@@ -26,6 +26,7 @@ Changelog
 
     - Bug fixes
         - Fix :func:`~twitchio.utils.setup_logging` breaking coloured formatting on `CRITICAL` logging level
+        - Fix :class:`~models.eventsub_.ChannelPointsAutoRedeemAdd` now accounts for attribute message in payload to be None
 
 
 3.1.0